Refactor DepositWithdraw page into component
