+++ src/components/DepositWithdraw/DepositWithdrawPage.css
@@ -1,0 +1,61 @@
+.deposit-withdraw-page {
+  padding: 20px;
+  max-width: 700px;
+  margin: 20px auto;
+}
+
+.tabs {
+  margin-bottom: 20px;
+  border-bottom: 1px solid var(--border);
+  display: flex;
+}
+
+.tab-button {
+  padding: 10px 15px;
+  cursor: pointer;
+  border: 1px solid transparent;
+  border-bottom: 2px solid transparent;
+  background-color: transparent;
+  color: var(--primary-text);
+  font-weight: normal;
+}
+
+.tab-button + .tab-button {
+  margin-left: 10px;
+}
+
+.tab-button.active {
+  border-bottom-color: var(--primary-blue);
+  color: var(--primary-blue);
+  font-weight: bold;
+}
+
+.coin-selection {
+  margin-bottom: 20px;
+  padding-bottom: 15px;
+  border-bottom: 1px solid var(--border);
+}
+
+.coin-button {
+  padding: 8px 12px;
+  margin: 5px;
+  cursor: pointer;
+  border: 1px solid var(--border);
+  background-color: var(--panel);
+  border-radius: 4px;
+}
+
+.coin-button.active {
+  border-color: var(--primary-blue);
+  background-color: #e7f3ff;
+}
+
+.select-coin-msg {
+  text-align: center;
+  color: var(--secondary-text);
+}
+
+.login-required {
+  padding: 20px;
+  text-align: center;
+}
